Use long IDs and update tests
Change ID types to primitive long in UserController and UserPreferencesRepository to align signatures. Update multiple unit tests to match API changes and new behaviors: add mocks (OrderService, OrderPDFService) where needed, adapt MailServiceTest to use MimeMessage and MailTemplateRenderer, adjust OrderServiceTest to return Optionals and mock user resolution, remove some ReflectionTestUtils settings in OrderPDFServiceTest, and update argument matchers (anyLong) and verifications in OrderStatusMailListenerTest and UserServiceTest. These updates ensure consistency of ID types and bring tests in line with the evolved service behavior.

+++ src/main/java/de/hs_esslingen/besy/controllers/UserController.java
@@ -37,7 +37,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserResponseDTO> getUser(@PathVariable("id") Integer id) {
+    public ResponseEntity<UserResponseDTO> getUser(@PathVariable("id") long id) {
         return ResponseEntity.ok(userResponseMapper.toDto(userService.getUserById(id)
                 .orElseThrow(() -> new NotFoundException("Benutzer mit id " + id + " nicht gefunden."))));
     }
